backend/requirements.txt
pypdf
astrapy
fastapi
google-genai
uvicorn